requirements.txt
datasets
imageio
imageio-ffmpeg
imagesize
imbalanced-learn
immutabledict
importlib_metadata
importlib_resources
imutils
inflect
iniconfig
ipyevents
isoduration
itsdangerous
jeepney
jieba
Jinja2
jiter
joblib
jsonpatch
jsonpickle
jsonpointer
jsonschema
jsonschema-specifications
kaggle
kagglehub
keras
keras-hub
keras-nlp
keyring
lazy_loader
linkify-it-py
llvmlite
locket
logical-unification
lxml
Mako
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
matplotlib-venn
mcp
mdurl
miniKanren
missingno
mistune
mizani
mkl
ml_dtypes
mlxtend
more-itertools
moviepy
mpmath
msgpack
multidict
multipledispatch
multiprocess
multitasking
murmurhash
music21
namex
narwhals
natsort
nbclassic
nbclient
nbconvert
nbformat
ndindex
nest-asyncio
networkx
nibabel
nltk
notebook
notebook_shim
numba
numba-cuda
numexpr
numpy
nvtx
oauth2client
oauthlib
omegaconf
openai
opencv-contrib-python
opencv-python
opencv-python-headless
openpyxl
pandas
pandas-datareader
pandas-gbq
pandas-stubs
pandocfilters
panel
param
parso
parsy
partd
patsy
peewee
peft
pexpect
pickleshare
pillow
platformdirs
plotly
plotnine
pluggy
plum-dispatch
ply
polars
pooch
portpicker
preshed
prettytable
proglog
progressbar2
prometheus_client
promise
prompt_toolkit
propcache
prophet
proto-plus
protobuf
psutil
psycopg2
psygnal
ptyprocess
py-cpuinfo
py4j
pyarrow
pyasn1
pyasn1_modules
pycairo
pycocotools
pycparser
pycryptodomex
pydantic
pydantic-settings
pydantic_core
pydata-google-auth
pydot
pydotplus
PyDrive2
pydub
pyerfa
pygame
pygit2
Pygments
PyJWT
pymc
pynndescent
pynvml
pyogrio
pyomo
PyOpenGL
pyOpenSSL
pyparsing
pyperclip
pyproj
pyproject_hooks
pyshp
PySocks
pyspark
pytensor
pytest
python-box
python-dateutil
python-dotenv
python-json-logger
python-louvain
python-multipart
python-slugify
python-snappy
python-utils
pytz
pyviz_comms
PyWavelets
PyYAML
pyzmq
ratelim
referencing
regex
requests
requests-oauthlib
requests-toolbelt
requirements-parser
rfc3339-validator
rfc3986-validator
rfc3987-syntax
rich
ruff
scikit-image
scikit-learn
scipy
scs
seaborn
simple-parsing
simplejson
simsimd
six
sklearn-pandas
slicer
smart_open
smmap
sniffio
snowballstemmer
sortedcontainers
soundfile
soupsieve
soxr
spacy
spacy-legacy
spacy-loggers
spanner-graph-notebook
srsly
sse-starlette
stanio
starlette
statsmodels
stringzilla
stumpy
sympy
tensorboard
tensorboard-data-server
tensorstore
tf-slim
tf_keras
tqdm
zipp